filter out snp position with alts in < 10 individuals
- This matches the workflow for GTEX (though they have a number of
other QC checks)
- It also solves the problem of singleton genotypes that happen to be
linked to an expression outlier being flagged as eQTLs
- This dramatically reduces the number of eQTLs found, so it’s possible
to lower the value threshold.
- I’ve also stopped filtering of FDR as I’m selecting values that
should mean that all results are FDR < 0.1 (further filtering is
possible in the shiny app).
- I tried this with value of 1e-4 for trans eQTLs and got 5 million at
FDR < 1e-10 and 370 million at FDR<0.1
- The 37GB file that this produces is not really manageable, so I’ve
dropped the value threshold for this back down to 1e-10
- I’m also filtering this dataset on FDR<1e-10 in case the unfiltered
dataset is still too unwieldy
- I’ve