rearrange rules
Previously, I was using the taxon in the final residualised counts file
to select the samples for the vcf, but that meant that the PCA results
that are included in the PEER analysis that produces the residualised
counts were inferred using samples that are not included in the final
analysis.

The new work-flow also means that the linkage to scz SNPs is determined
using only the samples that are used in the eQTL analysis.

Finally, it means that the VCF is filtered before import into Plink, so
I was able to remove the filtering from there, which should further
improve consistency