verify that genotyping and RNAseq sable IDs match
Requires mapping with HiSAT mapping, SNP calling and running GTcheck to
compare resulting VCF to genotyping file.

This is all somewhat unrelated to the main workflow, so I’ve abstracted
it to a separate snakefile

I’m unclear on how best to organise this. Mapping was originally done
for the Sex Diffs project, but I didn’t produce genotyping data for
that. This project uses the expression results from that one. Maybe the
most sensible place for this is the ASE project, which will require
genotyping data and mappings