use bed file from filter_counts for fastqtl
I had originally written an additional make_bed rule to cover the
residualised counts from PEER to bed format.

When we decided to use the raw counts and cofactors as input for
fastqtl, I kept the make_bed rule, which was now taking a bed file of
counts and creating a bed file of counts.