filter out GTEx eQTLs that we did not test
- this uses the the SNP ids from the overlapping_snps file rather
than the positions, so it's not necessary to go through the convert
to bed/liftover/sort/bedtools workflow for each tissue, only for the
snp map

- I benchmarked this agains bedtools and it's marginally slower and uses
more memory, but not enough to be problem. Once the liftover, etc is
factored in, this is way faster (tho I've already done the latter)

- As I write this, I'm realising that there's no easy way to use these
filtered files to filter our snp lists because the snp_ids are different.
I'm going to modify this rule to add our SNP ids to the GTEx files